requirements.txt
Django==4.2.7
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.1
seaborn==0.13.0
scikit-learn==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
Pillow==10.1.0
joblib==1.3.2
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8